Vista de TODOS por usuario

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import *as yup from 'yup';
@@ -29,23 +30,37 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if ( !user ) {
+        return NextResponse.json( 'NO autorizado', { status: 401 } )
+    }
+
+
     try {
         const { complete, description } = await postSchema.validate(await request.json());
-        const todo = await prisma.todo.create({ data: { description, complete}})
+        const todo = await prisma.todo.create({ data: { description, complete, userId: user.id}})
 
         return NextResponse.json(todo);
 
     } catch (error) {
       return NextResponse.json(error, { status: 400 })  
     }
+    
 
 }
 
 export async function DELETE(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if ( !user ) {
+        return NextResponse.json( 'NO autorizado', { status: 401 } )
+    }
+
     try {
         
-        const todo = await prisma.todo.deleteMany({ where: { complete: true}})
+        const todo = await prisma.todo.deleteMany({ where: { complete: true, userId: user.id}})
 
         return NextResponse.json('Borrados');
